pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "caldav>=2.1.2",
    "cryptography>=46.0.3",
    "fastapi>=0.123.5",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "httpx>=0.28.1",
    "icalendar>=6.3.2",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
